make benefit accordions TypeScript
