+++ src/Assets/questions.tsx
@@ -23,9 +23,9 @@ import signUpOptions from './signUpOptions.tsx';
 import healthInsuranceOptions from './healthInsuranceOptions.tsx';
 import acuteConditionOptions from './acuteConditionOptions';
 import { FormattedMessage } from 'react-intl';
-import type { QuestionNames, Question } from '../Types/Questions.ts';
+import type { QuestionName, Question } from '../Types/Questions.ts';
 
-const questions: Record<QuestionNames, Question> = {
+const questions: Record<QuestionName, Question> = {
   zipcode: {
     name: 'zipcode',
     question: <FormattedMessage id="questions.zipcode" defaultMessage="What is your zip code?" />,
