+++ src/Assets/stepDirectory.ts
@@ -1,8 +1,8 @@
 import questions from './questions';
-import type { QuestionNames } from '../Types/Questions';
+import type { QuestionName } from '../Types/Questions';
 
 export const startingQuestionNumber = 3;
-const defaultStepDirectory: QuestionNames[] = [
+const defaultStepDirectory: QuestionName[] = [
   'zipcode',
   'healthInsurance',
   //the hhSize and hhData have to be consecutive
@@ -15,7 +15,7 @@ const defaultStepDirectory: QuestionNames[] = [
   'referralSource',
   'signUpInfo',
 ];
-const referrerStepDirectory: QuestionNames[] = [
+const referrerStepDirectory: QuestionName[] = [
   'zipcode',
   'healthInsurance',
   //the hhSize and hhData have to be consecutive
@@ -37,7 +37,7 @@ export function getStepDirectory(referrer: string | undefined) {
   return defaultStepDirectory;
 }
 
-export function getStepNumber(name: QuestionNames, referrer: string | undefined) {
+export function getStepNumber(name: QuestionName, referrer: string | undefined) {
   const stepNumber = getStepDirectory(referrer).findIndex((question) => question === name);
 
   if (stepNumber === -1) {
