+++ src/Components/Confirmation/Confirmation.js
@@ -8,13 +8,13 @@ import frequencyOptions from '../../Assets/frequencyOptions';
 import expenseOptions from '../../Assets/expenseOptions';
 import healthInsuranceOptions from '../../Assets/healthInsuranceOptions.tsx';
 import acuteConditionOptions from '../../Assets/acuteConditionOptions';
-import cashAssistanceBenefits from '../../Assets/BenefitCategoryLists/cashAssistanceBenefits';
-import foodAndNutritionBenefits from '../../Assets/BenefitCategoryLists/foodAndNutritionBenefits';
-import childCareBenefits from '../../Assets/BenefitCategoryLists/childCareBenefits';
-import housingAndUtilities from '../../Assets/BenefitCategoryLists/housingAndUtilities';
-import transportationBenefits from '../../Assets/BenefitCategoryLists/transportationBenefits';
-import healthCareBenefits from '../../Assets/BenefitCategoryLists/healthCareBenefits';
-import taxCreditBenefits from '../../Assets/BenefitCategoryLists/taxCreditBenefits';
+import cashAssistanceBenefits from '../../Assets/BenefitCategoryLists/categories/cashAssistanceBenefits.tsx';
+import foodAndNutritionBenefits from '../../Assets/BenefitCategoryLists/categories/foodAndNutritionBenefits';
+import childCareBenefits from '../../Assets/BenefitCategoryLists/categories/childCareBenefits';
+import housingAndUtilities from '../../Assets/BenefitCategoryLists/categories/housingAndUtilities';
+import transportationBenefits from '../../Assets/BenefitCategoryLists/categories/transportationBenefits';
+import healthCareBenefits from '../../Assets/BenefitCategoryLists/categories/healthCareBenefits';
+import taxCreditBenefits from '../../Assets/BenefitCategoryLists/categories/taxCreditBenefits';
 import { getStepDirectory, getStepNumber, startingQuestionNumber } from '../../Assets/stepDirectory';
 import { useContext, useEffect } from 'react';
 import { Context } from '../Wrapper/Wrapper.tsx';
